+++ src/hotspot/share/runtime/signature.hpp
@@ -580,13 +580,13 @@ class SigEntry {
   // TODO improve these comments
   BasicType _bt;
   int _offset;      // Offset of the corresponding field in it's value class holder for scalarized arguments (-1 otherwise). Used for packing and unpacking
-  int _sort_offset; // Offset used for sorting
+  float _sort_offset; // Offset used for sorting
   Symbol* _symbol;  // For printing
 
   SigEntry()
     : _bt(T_ILLEGAL), _offset(-1), _sort_offset(-1), _symbol(NULL) {}
 
-  SigEntry(BasicType bt, int offset = -1, int sort_offset = -1, Symbol* symbol = nullptr)
+  SigEntry(BasicType bt, int offset = -1, float sort_offset = -1, Symbol* symbol = nullptr)
     : _bt(bt), _offset(offset), _sort_offset(sort_offset), _symbol(symbol) {}
 
   static int compare(SigEntry* e1, SigEntry* e2) {
@@ -614,7 +614,7 @@ class SigEntry {
     ShouldNotReachHere();
     return 0;
   }
-  static void add_entry(GrowableArray<SigEntry>* sig, BasicType bt, Symbol* symbol = nullptr, int offset = -1, int sort_offset = -1);
+  static void add_entry(GrowableArray<SigEntry>* sig, BasicType bt, Symbol* symbol = nullptr, int offset = -1, float sort_offset = -1);
   static bool skip_value_delimiters(const GrowableArray<SigEntry>* sig, int i);
   static int fill_sig_bt(const GrowableArray<SigEntry>* sig, BasicType* sig_bt);
   static TempNewSymbol create_symbol(const GrowableArray<SigEntry>* sig);
